Rework grid as an object
Assuming a square grid of dimension 'side', grid coordinates are
now integers in the range 0..side²-1 (which allows using an array
instead of a dictionary when associating a value with each cell).

The location generation algorithm has been reworked ; instead of
throwing values at a random angle from the center, it now builds a
tree of locations at a distance of exactly three from each other
(and no closer).