Implement hex cell selection
Now, clicking on a cell "picks" it, making it selected and centering it.
Clicking anywhere while a cell is selected unselects it but does not
move the map view.

This is a temporary state, as there is no long-term plan to allow
the selection of non-location cells (and there are graphical glitches
that appear when it is allowed).