Fix z-index of location names on the map
When a filter is applied (for instance, in the fog of war), the entire
tile is combined, which prevents the location-name element from having
a distinct z-index.

This commit instead moves the hex rendering (and filters) from the
hex element's root (and parent of the location-name element) to
a sibling of the location-name element, which allows the two to be
kept separate and have different z-index values.