Rework initial occupation extraction
The occupation module now has a table describing ease of recruitment,
which is used by the world generation module to pick an initial
occupation based on the location type.

The location-type enum has been introduced to keep track of the
location types (instead of relying on cell kinds).

Renamed "Temple" to "Magic Academy" for consistency.