Rework "currently selected" system
It is now managed at the root level in Screen.tsx, which means that
moving agents, when they remain selected, are followed on the map
automatically.

Also, clicking on a selectable entity on the map when something else
is selected, instead of just clearing the selection, actually selects
the selectable entity.