Produce a binary pack and generate <style> at runtime
Previously, assets were included in the CSS file as data-urls, which
were base64-encoded and therefore unnecessarily large, and also
made it hard to edit the style rules from a browser's developer
tools (since the properties were huge).

Now, all the assets are packed together in a binary blob, which is
then downloaded and sliced into individual blobs on the client side,
and included in a <style> using `URL.createObjectURL()`.

This cuts the assets size from 4.645KB to 3.472KB (though the gzipped
size is the same in both cases, arguably decoding base64 data-url is
slower than createObjectURL).