Implement loading/saving
This is a big commit, but it needed to be done sooner rather than
later, because as the model complexity increases it becomes harder
to add serialization on top of it.

The point of using an explicit serializer instead of just JSON-ing
everything is because some objects have fields that require complex
initialization (such as shared references) and others contain large
amounts of Float32 or Uint16 data. Using an explicit serializer also
ensures that all fields are accounted for.