Rework district generation
Now the number of districts is based on the type of the location and the
population count, and the population is then spread among the districts
using a weight modifier (so that greens or mines receive less population
than residential districts).

The "workcamp" location kind is now split into three kinds: ironworks,
lumber camp, and mine. They each get specific districts.

Added a name generator for districts.